+++ doc/changelog.md
@@ -8,6 +8,7 @@
 - Added `base` library support for UTF-8.
 - Added `base` library support for `atoi`-like string conversion.
 - Added `endwhile` syntax.
+- Added `example/convert-to-nia`.
 - Added `example/imageviewer`.
 - Added `example/jsonptr`.
 - Added `std/bmp`.